Overwrite files when instructed to overwrite files

+++ broadcast-transcoder/src/main/java/dk/statsbiblioteket/broadcasttranscoder/processors/FinalMediaFileRenamerProcessor.java
@@ -9,6 +9,7 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.StandardCopyOption;
 
 /**
  *
@@ -22,17 +23,16 @@ public class FinalMediaFileRenamerProcessor extends ProcessorChainElement {
         final File finalFile = FileUtils.getFinalMediaOutputFile(request, context);
         logger.debug("Renaming " + tempFile.getAbsolutePath() + " to " + finalFile.getAbsolutePath());
         try {
-            Files.move(tempFile.toPath(),finalFile.toPath());
+            if (context.isOverwrite()){
+                Files.move(tempFile.toPath(), finalFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            } else {
+                Files.move(tempFile.toPath(), finalFile.toPath());
+            }
         } catch (IOException e) {
-            final String s = "Failed to rename " + tempFile.getAbsolutePath() + " to " + finalFile.getAbsolutePath();
+            final String s = "Failed to rename '" + tempFile.getAbsolutePath() + "' to '" + finalFile.getAbsolutePath()+"' . Overwrite is set to "+context.isOverwrite();
             logger.error(s,e);
             throw new ProcessorException(s,e);
         }
-//        if (!tempFile.renameTo(finalFile)) {
-//            final String s = "Failed to rename " + tempFile.getAbsolutePath() + " to " + finalFile.getAbsolutePath();
-//            logger.error(s);
-//            throw new ProcessorException(s);
-//        }
     }
